+++ Common/Authentication/OpenShockAuthSchemas.cs
@@ -7,4 +7,7 @@ public static class OpenShockAuthSchemas
     public const string HubToken = "HubToken";
 
     public const string UserSessionApiTokenCombo = $"{UserSessionCookie},{ApiToken}";
+
+    public const string DiscordScheme = "discord";
+    public static readonly string[] OAuth2Schemes = [DiscordScheme];
 }
