+++ Common/Constants/AuthConstants.cs
@@ -7,9 +7,6 @@ public static class AuthConstants
     public const string ApiTokenHeaderName = "OpenShockToken";
     public const string HubTokenHeaderName = "DeviceToken";
 
-    public const string DiscordScheme = "discord";
-    public static readonly string[] OAuth2Schemes = [DiscordScheme];
-
     public const int GeneratedTokenLength = 32;
     public const int ApiTokenLength = 64;
 }
